viewer node callables' names are on module as well
In other words, the callables can now be found on the module using
their names, not only using the 'main_callable' variable.

This is important because the feature of exporting node layouts
as python scripts rely on the name of the callables to find them
on the node scripts, not on the 'main_callable' variable.